chore: dedupe running publish.test tests twice
The tail.test suite was importing a helper from publish.test, which was also declaring all the tests in that suite. This fix extracts those helpers so they can be imported and used cleanly. Shaves a few seconds off when running all tests too.
