+++ packages/wrangler/src/__tests__/helpers/mock-known-routes.ts
@@ -1,0 +1,7 @@
+import { setMockResponse } from "./mock-cfetch";
+
+export function mockCollectKnownRoutesRequest(
+	routes: { pattern: string; script: string }[]
+) {
+	setMockResponse(`/zones/:zoneId/workers/routes`, "GET", () => routes);
+}
